refactor(backend): create GenerateDataMartMetadataOrchestratorService, add DataMartAiHelperGeneratedEvent
