+++ apps/backend/src/data-marts/ai-insights/data-mart-insights.types.ts
@@ -11,7 +11,7 @@ export type DataMartAdditionalParams = {
 };
 
 export interface ConsumptionContext {
-  contextType: 'INSIGHT' | 'REPORT' | 'DATA_MART_HELPER';
+  contextType: 'INSIGHT' | 'REPORT';
   contextId: string;
   contextTitle: string;
   dataMart: DataMart;
