+++ apps/backend/src/data-marts/ai-insights/facades/ai-insights.facade.impl.ts
@@ -4,22 +4,16 @@ import { AiInsightsFacade } from './ai-insights.facade';
 import {
   AnswerPromptRequest,
   AnswerPromptResponse,
-  DataMartMetadataScope,
   GenerateDataMartMetadataRequest,
   GenerateDataMartMetadataResponse,
   GenerateInsightRequest,
   GenerateInsightResponse,
-  QueryRow,
   SharedAgentContext,
 } from '../ai-insights-types';
 import { AiInsightsOrchestratorService } from '../ai-insight-orchestrator.service';
-import {
-  ConsumptionContext,
-  DataMartPromptMetaEntry,
-  PromptAnswer,
-} from '../data-mart-insights.types';
+import { DataMartPromptMetaEntry, PromptAnswer } from '../data-mart-insights.types';
 import { GenerateInsightAgent } from '../agent/generate-insight.agent';
-import { GenerateDataMartMetadataAgent } from '../agent/generate-data-mart-metadata.agent';
+import { GenerateDataMartMetadataOrchestratorService } from '../generate-data-mart-metadata.orchestrator.service';
 import { AI_CHAT_PROVIDER } from '../../../common/ai-insights/services/ai-chat-provider.token';
 import { AiChatProvider } from '../../../common/ai-insights/agent/ai-core';
 import { ToolRegistry } from '../../../common/ai-insights/agent/tool-registry';
@@ -29,15 +23,6 @@ import {
   measureExecutionTime,
   toMeasuredExecutionBaseResult,
 } from '../../../common/utils/measure-execution-time';
-import { DataMartService } from '../../services/data-mart.service';
-import { DataMartSqlTableService } from '../../services/data-mart-sql-table.service';
-import { DataMartDefinitionValidatorFacade } from '../../data-storage-types/facades/data-mart-definition-validator-facade.service';
-import { ConsumptionTrackingService } from '../../services/consumption-tracking.service';
-import { DataMartDefinitionType } from '../../enums/data-mart-definition-type.enum';
-import { BusinessViolationException } from '../../../common/exceptions/business-violation.exception';
-import { getPromptTotalUsage } from '../utils/compute-model-usage';
-
-const METADATA_SAMPLE_ROW_LIMIT = 30;
 
 @Injectable()
 export class AiInsightsFacadeImpl implements AiInsightsFacade {
@@ -46,11 +31,7 @@ export class AiInsightsFacadeImpl implements AiInsightsFacade {
   constructor(
     private readonly aiInsightsAgentService: AiInsightsOrchestratorService,
     private readonly generateInsightAgent: GenerateInsightAgent,
-    private readonly generateDataMartMetadataAgent: GenerateDataMartMetadataAgent,
-    private readonly dataMartService: DataMartService,
-    private readonly dataMartSqlTableService: DataMartSqlTableService,
-    private readonly definitionValidatorFacade: DataMartDefinitionValidatorFacade,
-    private readonly consumptionTracker: ConsumptionTrackingService,
+    private readonly generateDataMartMetadataOrchestratorService: GenerateDataMartMetadataOrchestratorService,
     @Inject(AI_CHAT_PROVIDER)
     private readonly aiProvider: AiChatProvider,
     private readonly toolRegistry: ToolRegistry
@@ -162,146 +143,19 @@ export class AiInsightsFacadeImpl implements AiInsightsFacade {
   async generateDataMartMetadata(
     request: GenerateDataMartMetadataRequest
   ): Promise<GenerateDataMartMetadataResponse> {
-    this.logger.log(
-      `Generating data mart metadata for ${request.dataMartId} (scope=${request.scope}, useSample=${request.useSample})`
-    );
-
-    const dataMart = await this.dataMartService.getByIdAndProjectId(
-      request.dataMartId,
-      request.projectId
-    );
-
-    if (dataMart.definitionType === DataMartDefinitionType.CONNECTOR) {
-      throw new BusinessViolationException(
-        'AI metadata generation is not supported for connector-based data marts'
-      );
-    }
-
-    await this.definitionValidatorFacade.checkIsValid(dataMart);
-
-    const schema = dataMart.schema;
-    if (!schema) {
-      throw new BusinessViolationException(
-        'Data mart has no output schema yet. Actualize the schema before requesting AI metadata.'
-      );
-    }
-
-    if (this.requiresFieldName(request.scope) && !request.fieldName?.trim()) {
-      throw new BusinessViolationException(`fieldName is required for scope "${request.scope}"`);
-    }
-
-    let sampleColumns: string[] | null = null;
-    let sampleRows: QueryRow[] | null = null;
-    if (request.useSample) {
-      const sample = await this.dataMartSqlTableService.executeSqlToTable(dataMart, undefined, {
-        limit: METADATA_SAMPLE_ROW_LIMIT,
-      });
-      sampleColumns = sample.columns;
-      sampleRows = sample.rows.map(row => this.toQueryRow(sample.columns, row));
-    }
-
-    const sharedContext: SharedAgentContext = {
-      aiProvider: this.aiProvider,
-      toolRegistry: this.toolRegistry,
-      budgets: {},
-      telemetry: {
-        llmCalls: [],
-        toolCalls: [],
-        messageHistory: [],
-      },
-      projectId: request.projectId,
-      dataMartId: request.dataMartId,
-    };
-
     const measured = await measureExecutionTime(
-      () =>
-        this.generateDataMartMetadataAgent.run(
-          {
-            scope: request.scope,
-            dataMartTitle: dataMart.title ?? null,
-            dataMartDescription: dataMart.description ?? null,
-            schema,
-            sampleColumns,
-            sampleRows,
-            fieldName: request.fieldName,
-          },
-          sharedContext
-        ),
+      () => this.generateDataMartMetadataOrchestratorService.run(request),
       {
-        onMeasured: measuredExecution => {
+        onMeasured: m => {
           this.logger.log('AiGenerateDataMartMetadataTime', {
             projectId: request.projectId,
             dataMartId: request.dataMartId,
             scope: request.scope,
-            measure: toMeasuredExecutionBaseResult(measuredExecution),
+            measure: toMeasuredExecutionBaseResult(m),
           });
         },
       }
     );
-
-    const aiResult = measured.result;
-
-    const consumptionContext: ConsumptionContext = {
-      contextType: 'DATA_MART_HELPER',
-      contextId: dataMart.id,
-      contextTitle: dataMart.title,
-      dataMart,
-    };
-    try {
-      const totalTokens = getPromptTotalUsage(sharedContext.telemetry.llmCalls).totalTokens;
-      void this.consumptionTracker.registerAiProcessRunConsumption(totalTokens, consumptionContext);
-    } catch (error) {
-      this.logger.error('Failed to register AI metadata helper consumption', error);
-    }
-
-    return this.shapeResponseToScope(aiResult, request);
-  }
-
-  private requiresFieldName(scope: DataMartMetadataScope): boolean {
-    return (
-      scope === DataMartMetadataScope.FIELD_ALIAS ||
-      scope === DataMartMetadataScope.FIELD_DESCRIPTION
-    );
-  }
-
-  private toQueryRow(columns: string[], row: unknown[]): QueryRow {
-    const result: QueryRow = {};
-    for (let i = 0; i < columns.length; i++) {
-      result[columns[i]] = row[i];
-    }
-    return result;
-  }
-
-  private shapeResponseToScope(
-    aiResult: GenerateDataMartMetadataResponse,
-    request: GenerateDataMartMetadataRequest
-  ): GenerateDataMartMetadataResponse {
-    const { scope, fieldName } = request;
-
-    switch (scope) {
-      case DataMartMetadataScope.TITLE:
-        return { title: aiResult.title };
-      case DataMartMetadataScope.DESCRIPTION:
-        return { description: aiResult.description };
-      case DataMartMetadataScope.FIELD_ALIAS: {
-        const match = aiResult.fields?.find(f => f.name === fieldName);
-        return { fields: match ? [{ name: match.name, alias: match.alias }] : [] };
-      }
-      case DataMartMetadataScope.FIELD_DESCRIPTION: {
-        const match = aiResult.fields?.find(f => f.name === fieldName);
-        return {
-          fields: match ? [{ name: match.name, description: match.description }] : [],
-        };
-      }
-      case DataMartMetadataScope.ALL_FIELD_ALIASES:
-        return {
-          fields: aiResult.fields?.map(f => ({ name: f.name, alias: f.alias })) ?? [],
-        };
-      case DataMartMetadataScope.ALL_FIELD_DESCRIPTIONS:
-      default:
-        return {
-          fields: aiResult.fields?.map(f => ({ name: f.name, description: f.description })) ?? [],
-        };
-    }
+    return measured.result;
   }
 }
