+++ apps/backend/src/data-marts/use-cases/generate-data-mart-metadata.service.ts
@@ -12,6 +12,8 @@ import { GenerateDataMartMetadataCommand } from '../dto/domain/generate-data-mar
 import { DataMartService } from '../services/data-mart.service';
 import { AccessDecisionService, Action, EntityType } from '../services/access-decision';
 import { AiInsightsConfigService } from '../../common/ai-insights/services/ai-insights-config.service';
+import { OwoxEventDispatcher } from '../../common/event-dispatcher/owox-event-dispatcher';
+import { DataMartAiHelperGeneratedEvent } from '../events/data-mart-ai-helper-generated.event';
 
 @Injectable()
 export class GenerateDataMartMetadataService {
@@ -21,6 +23,7 @@ export class GenerateDataMartMetadataService {
     private readonly dataMartService: DataMartService,
     private readonly accessDecisionService: AccessDecisionService,
     private readonly aiInsightsConfig: AiInsightsConfigService,
+    private readonly eventDispatcher: OwoxEventDispatcher,
     @Inject(AI_INSIGHTS_FACADE)
     private readonly aiInsightsFacade: AiInsightsFacade
   ) {}
@@ -52,12 +55,31 @@ export class GenerateDataMartMetadataService {
       `Generating metadata for data mart ${command.id} (scope=${command.scope}, useSample=${command.useSample})`
     );
 
-    return this.aiInsightsFacade.generateDataMartMetadata({
+    const result = await this.aiInsightsFacade.generateDataMartMetadata({
       projectId: command.projectId,
       dataMartId: command.id,
       scope: command.scope,
       useSample: command.useSample,
       fieldName: command.fieldName,
     });
+
+    // Fire-and-forget analytics — transport failures must not break the user response.
+    try {
+      await this.eventDispatcher.publishExternal(
+        new DataMartAiHelperGeneratedEvent({
+          projectId: command.projectId,
+          dataMartId: command.id,
+          userId: command.userId,
+          scope: command.scope,
+        })
+      );
+    } catch (error) {
+      this.logger.error(
+        `Failed to publish DataMartAiHelperGeneratedEvent (scope=${command.scope})`,
+        error
+      );
+    }
+
+    return result;
   }
 }
